Share tvOS Skyline feed across Home and libraries
- Add a shared Skyline section feed for tvOS landing rows
- Replace collection fan cards with poster-grid collection cards
- Default library scope entry to Recommended
